When we restore snapshot on a fail project do not crash
Fix #800

+++ gns3server/controller/snapshot.py
@@ -80,7 +80,8 @@ class Snapshot:
         # We don't send close notif to clients because the close / open dance is purely internal
         yield from self._project.close(ignore_notification=True)
         self._project.controller.notification.emit("snapshot.restored", self.__json__())
-        shutil.rmtree(os.path.join(self._project.path, "project-files"))
+        if os.path.exists(os.path.join(self._project.path, "project-files")):
+            shutil.rmtree(os.path.join(self._project.path, "project-files"))
         with open(self._path, "rb") as f:
             project = yield from import_project(self._project.controller, self._project.id, f, location=self._project.path)
         yield from project.open()
